+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 - Field `label` added to the template tests definitions. (@mihaitodor)
 - Metadata field `label` can now be utilized within a template's `mapping` field to access the label that is associated with the template instantiation in a config.
 
+### Changed
+
+- The `branch` processor no longer emits an entry in the log at error level when the child processors throw errors. (@mihaitodor)
+
 ## 4.42.0 - 2024-11-29
 
 ### Added